# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " This is an API developeded for Althaia",
        "termsOfService": "http://www.althaia.com.br/",
        "contact": {
            "email": "juracabral@yahoo.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "System Authentication",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Exit the System",
                "responses": {
                    "200": {
                        "description": "successful logged"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "description": "Returns list of products",
                "operationId": "App\\Http\\Controllers\\Api\\ProductsController::index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "description": "Create a new Product",
                "operationId": "App\\Http\\Controllers\\Api\\ProductsController::store",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "description": "Return a product",
                "operationId": "App\\Http\\Controllers\\Api\\ProductsController::show",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "description": "Edit a Product",
                "operationId": "App\\Http\\Controllers\\Api\\ProductsController::update",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "description": "Delete a product",
                "operationId": "App\\Http\\Controllers\\Api\\ProductsController::destroy",
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    }
}